Backend/requirements.txt
pydantic
requests
python-dotenv
fastapi
scikit-learn
uvicorn
asyncio
pypdf2
nest_asyncio
pillow
transformers
torch
keras
jax
jaxlib
ultralytics
python-multipart
PyPDF2
openai
python-dotenv